Format datetime as ISO 8601 with local timestamp
Prior to this change the datetime fields were stored in UTC time.
With this change datetime fiels are formated according to ISO 8601, with
US/Central timezone explictly indicated in the field value. E.g. 2016-06-15T10:00:00-05:00. The datets also have a TIME field, which was and contiues to be stored as hours, minutes, and seconds, in local time. Although we do not explicity inidicate the the TIME field is in local time, it is obvious thanks to the corresponding datetime field which does indicate the time